+++ src/app/admin/users/user-list/user-list-item/user-list-item.component.ts
@@ -14,14 +14,6 @@ import { SessionService } from '../../../../session.service'
 import { HandleSubscription } from 'common/handle-subscription'
 import { CODE, CRYPTO } from 'common/utilities/consts'
 import { User } from 'models/user.model'
-import { ShowDialogOnError } from 'store/common/common.actions'
-import { MatDialog } from '@angular/material/dialog'
-import { BanUserDialogComponent } from 'admin/dialog/ban-user-dialog/ban-user-dialog.component'
-import { DeleteUserDialogComponent } from 'admin/dialog/delete-user-dialog/delete-user-dialog.component'
-import {
-  UserConfirmResponseDialogComponent
-} from 'common/dialog/user-confirm-response-dialog/user-confirm-response-dialog.component'
-import { BanUser, DeleteUser, UnbanUser } from 'store/admin/admin.actions'
 import { ServerOptionsService } from 'common/server-options.service'
 
 
@@ -39,7 +31,6 @@ export class UserListItemComponent extends HandleSubscription implements OnInit 
   code: string = CODE
   calculateFunds: boolean
   faIconImpersonation = faUserSecret
-  isSaving: boolean = false;
 
   constructor (
     private adminService: AdminService,
@@ -48,7 +39,6 @@ export class UserListItemComponent extends HandleSubscription implements OnInit 
     private impersonationService: ImpersonationService,
     private sessionService: SessionService,
     private router: Router,
-    private dialog: MatDialog,
   ) {
     super()
   }
@@ -57,74 +47,6 @@ export class UserListItemComponent extends HandleSubscription implements OnInit 
     this.loggedUser = this.sessionService.getUser()
     const options = this.serverOptionsService.getOptions()
     this.calculateFunds = options.displayCurrency !== options.appCurrency
-  }
-
-  handleConfirmation (): void {
-    this.isSaving = true;
-    this.adminService.confirmUser(this.user.id).subscribe(
-      (user) => {
-        this.user = {
-          ...this.user,
-          ...user,
-        }
-        this.isSaving = false;
-      },
-      () => {
-        this.isSaving = false;
-        this.store.dispatch(new ShowDialogOnError(''));
-      }
-    )
-  }
-
-  handleSwitchingToModerator (): void {
-    this.isSaving = true;
-    this.adminService.switchToModerator(this.user.id).subscribe(
-      (user) => {
-        this.user = {
-          ...this.user,
-          ...user,
-        }
-        this.isSaving = false;
-      },
-      () => {
-        this.isSaving = false;
-        this.store.dispatch(new ShowDialogOnError(''));
-      }
-    )
-  }
-
-  handleSwitchingToAgency (): void {
-    this.isSaving = true;
-    this.adminService.switchToAgency(this.user.id).subscribe(
-      (user) => {
-        this.user = {
-          ...this.user,
-          ...user,
-        }
-        this.isSaving = false;
-      },
-      () => {
-        this.isSaving = false;
-        this.store.dispatch(new ShowDialogOnError(''));
-      }
-    )
-  }
-
-  handleSwitchingToRegular (): void {
-    this.isSaving = true;
-    this.adminService.switchToRegular(this.user.id).subscribe(
-      (user) => {
-        this.user = {
-          ...this.user,
-          ...user,
-        }
-        this.isSaving = false;
-      },
-      () => {
-        this.isSaving = false;
-        this.store.dispatch(new ShowDialogOnError(''));
-      }
-    )
   }
 
   handleImpersonating (): void {
@@ -140,23 +62,6 @@ export class UserListItemComponent extends HandleSubscription implements OnInit 
           this.sessionService.setAccountTypeChoice(SessionService.ACCOUNT_TYPE_ADVERTISER)
         }
       },
-    )
-  }
-
-  handleUserRightsChange(operation: string): void {
-    this.isSaving = true;
-    this.adminService.changeUserRights(this.user.id, operation).subscribe(
-      (user) => {
-        this.user = {
-          ...this.user,
-          ...user,
-        }
-        this.isSaving = false;
-      },
-      () => {
-        this.isSaving = false;
-        this.store.dispatch(new ShowDialogOnError(''));
-      }
     )
   }
 
@@ -211,44 +116,5 @@ export class UserListItemComponent extends HandleSubscription implements OnInit 
     else {
       return 'User has no advertiser nor publisher role'
     }
-  }
-
-  showBanConfirmationDialog(){
-    this.dialog.open(BanUserDialogComponent, {
-      autoFocus: false,
-      data: this.user
-    })
-      .afterClosed()
-      .subscribe(result => {
-        if(result) {
-          this.store.dispatch(new BanUser({ id: this.user.id, reason: result }))
-        }
-      })
-  }
-
-  showUnbanConfirmationDialog(){
-    this.dialog.open(UserConfirmResponseDialogComponent, {
-      data: {
-        message: `Do you want unban user ${this.user.email}?`
-      }
-    })
-      .afterClosed()
-      .subscribe(result => {
-        if(result){
-          this.store.dispatch(new UnbanUser(this.user.id))
-        }
-      })
-  }
-
-  showDeleteConfirmationDialog(){
-    this.dialog.open(DeleteUserDialogComponent, {
-      data: this.user.email
-    })
-      .afterClosed()
-      .subscribe(result => {
-        if(result){
-          this.store.dispatch(new DeleteUser(this.user.id))
-        }
-      })
   }
 }